+++ docs/modules/python/opendevin/core/schema/observation.md
@@ -1,0 +1,39 @@
+---
+sidebar_label: observation
+title: opendevin.core.schema.observation
+---
+
+## ObservationTypeSchema Objects
+
+```python
+class ObservationTypeSchema(BaseModel)
+```
+
+#### READ
+
+The content of a file
+
+#### BROWSE
+
+The HTML content of a URL
+
+#### RUN
+
+The output of a command
+
+#### RUN\_IPYTHON
+
+Runs a IPython cell.
+
+#### RECALL
+
+The result of a search
+
+#### CHAT
+
+A message from the user
+
+#### DELEGATE
+
+The result of a task delegated to another agent
+
