+++ opendevin/core/schema/__init__.py
@@ -3,4 +3,10 @@ from .config import ConfigType
 from .observation import ObservationType
 from .task import TaskState, TaskStateAction
 
-__all__ = ['ActionType', 'ObservationType', 'ConfigType', 'TaskState', 'TaskStateAction']
+__all__ = [
+    'ActionType',
+    'ObservationType',
+    'ConfigType',
+    'TaskState',
+    'TaskStateAction',
+]
