+++ opendevin/events/observation/recall.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from typing import List
 
-from opendevin.schema import ObservationType
+from opendevin.core.schema import ObservationType
 
 from .observation import Observation
