+++ opendevin/runtime/sandbox/docker/exec_box.py
@@ -14,10 +14,10 @@ import docker
 from opendevin.core import config
 from opendevin.core.exceptions import SandboxInvalidBackgroundCommandError
 from opendevin.core.logger import opendevin_logger as logger
+from opendevin.core.schema import ConfigType
 from opendevin.runtime.sandbox.docker.process import DockerProcess
 from opendevin.runtime.sandbox.process import Process
 from opendevin.runtime.sandbox.sandbox import Sandbox
-from opendevin.schema import ConfigType
 
 InputType = namedtuple('InputType', ['content'])
 OutputType = namedtuple('OutputType', ['content'])
